refactor(iocuak): rename task module to be createInstanceTask
